Also rename support file, from fake_operations.rb to env_path.rb
